requirements.txt
fastapi
uvicorn[standard]
whoosh
python-docx
pypdf2 # ou pdfplumber
Markdown
pydantic